requirements.txt
prometheus-client
pydantic-settings>2
click
# Opensearch
opensearch-py[async]>=3.0.0
# Note this is a fork from the original https://github.com/dpkp/kafka-python which will be preferred in the future.
kafka-python-ng
# Additional compression options for kafka
python-snappy
lz4
# minio
minio
# Azure
azure-storage-blob
azure-identity
# redis
redis>5.1
# auth
Authlib
cachetools
asyncache
httpx
# Azul Client for probing
azul-client
